refactor: separate the enums from the types file
