+++ packages/api-postgres/src/auth/__tests__/auth.acceptance-test.ts
@@ -1,7 +1,7 @@
 import { createPostgresConnection, closeConnection } from '../../config/database';
 import TestClient from '../../testUtils/TestClient';
 import TestUtils from '../../testUtils/TestUtilities';
-import { Program } from '../../common/types';
+import { Program } from '../../common/types/enums';
 
 const expectedUserShape = {
     id: expect.any(String),
