+++ packages/api-postgres/src/common/types/enums.ts
@@ -1,0 +1,58 @@
+export enum UserRole {
+    ADMIN = 'admin',
+    APPROVER = 'approver',
+    REQUESTOR = 'requestor',
+    VOLUNTEER = 'volunteer',
+    UNKNOWN = 'unknown'
+}
+
+export enum Program {
+    HOUSING = 'housing first',
+    INTEGRATED = 'integrated services',
+    SURVIVAL = 'survival services',
+    HEALTH = 'health services',
+    EMPLOYMENT = 'employment services',
+    RESEARCH_INNOVATION = 'research_innovation',
+    UNKNOWN = 'unknown'
+}
+
+export enum ItemCategory {
+    CLOTHING = 'clothing',
+    ENGAGEMENT = 'engagement',
+    HOUSEHOLD = 'household',
+    PERSONAL_HYGIENE = 'personal hygiene',
+    PET = 'pet',
+    TICKET = 'ticket',
+    OTHER = 'other'
+}
+
+export enum ItemPriority {
+    URGENT = 'urgent',
+    STANDARD = 'standard'
+}
+
+export enum ItemStatus {
+    ACTIVE = 'active',
+    APPROVED = 'approved',
+    DENIED = 'denied',
+    FUFILLED = 'fulfilled',
+    WISHLIST = 'wishlist',
+    ARCHIVED = 'archived'
+}
+
+export enum HouseLocation {
+    EASTLAKE = 'eastlake',
+    AURORA_HOUSE = 'aurora house',
+    CANADAY_HOUSE = 'canaday house',
+    CLEMENT_PLACE = 'clement place',
+    COTTAGE_GROVE_COMMONS = 'cottage grove commons',
+    ESTELLE = 'estelle',
+    EVANS_HOUSE = 'evans house',
+    INTERBAY_PLACE = 'interbay place',
+    KERNER_SCOTT_HOUSE = 'kerner-scott house',
+    KEYS = 'keys',
+    LYON_BUILDING = 'lyon building',
+    MORRISON = 'morrison',
+    RAINIER_HOUSE = 'rainier house',
+    UNION_HOTEL = 'union hotel'
+}
