+++ packages/api-postgres/src/common/types/types.ts
@@ -1,20 +1,5 @@
-export enum UserRole {
-    ADMIN = 'admin',
-    APPROVER = 'approver',
-    REQUESTOR = 'requestor',
-    VOLUNTEER = 'volunteer',
-    UNKNOWN = 'unknown'
-}
+import { HouseLocation, ItemCategory, ItemPriority, ItemStatus, Program, UserRole } from './enums';
 
-export enum Program {
-    HOUSING = 'housing first',
-    INTEGRATED = 'integrated services',
-    SURVIVAL = 'survival services',
-    HEALTH = 'health services',
-    EMPLOYMENT = 'employment services',
-    RESEARCH_INNOVATION = 'research_innovation',
-    UNKNOWN = 'unknown'
-}
 export interface UserFields {
     firstName: string;
     lastName: string;
@@ -28,47 +13,6 @@ export interface UserFields {
 }
 
 export type UpdatableUserFields = Partial<Omit<UserFields, 'password' | 'roles'>>;
-
-export enum ItemCategory {
-    CLOTHING = 'clothing',
-    ENGAGEMENT = 'engagement',
-    HOUSEHOLD = 'household',
-    PERSONAL_HYGIENE = 'personal hygiene',
-    PET = 'pet',
-    TICKET = 'ticket',
-    OTHER = 'other'
-}
-
-export enum ItemPriority {
-    URGENT = 'urgent',
-    STANDARD = 'standard'
-}
-
-export enum ItemStatus {
-    ACTIVE = 'active',
-    APPROVED = 'approved',
-    DENIED = 'denied',
-    FUFILLED = 'fulfilled',
-    WISHLIST = 'wishlist',
-    ARCHIVED = 'archived'
-}
-
-export enum HouseLocation {
-    EASTLAKE = 'eastlake',
-    AURORA_HOUSE = 'aurora house',
-    CANADAY_HOUSE = 'canaday house',
-    CLEMENT_PLACE = 'clement place',
-    COTTAGE_GROVE_COMMONS = 'cottage grove commons',
-    ESTELLE = 'estelle',
-    EVANS_HOUSE = 'evans house',
-    INTERBAY_PLACE = 'interbay place',
-    KERNER_SCOTT_HOUSE = 'kerner-scott house',
-    KEYS = 'keys',
-    LYON_BUILDING = 'lyon building',
-    MORRISON = 'morrison',
-    RAINIER_HOUSE = 'rainier house',
-    UNION_HOTEL = 'union hotel'
-}
 
 // Sizes
 export const availableShirtOrCoatSizes = [
