+++ packages/api-postgres/src/item/ItemService.ts
@@ -1,5 +1,5 @@
 import Item from '../entity/Item';
-import { ItemFields, UpdatableItemFields } from '../common/types';
+import { ItemFields, UpdatableItemFields } from '../common/types/types';
 import UserService from '../user/UserService';
 import NoteService from '../note/NoteService';
